+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * Added capability to change the PrimarySMTPAddress of a Shared Mailbox
 * SPOExternalUserExpireInDays
   * Added the External User Expiration setting in the config.
+* ExternalCrossTenantTrust
+  * Added the automatic user consent settings for Inbound and Outbound allowed
 * MISC
   * Updated AAD, EXO and Teams settings file to describe required roles.
   * Added a new personas documentation page to describe the targeted personas for the project.