+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change log for Microsoft365DSC
 
 # UNRELEASED
+
 * AADGroup
   * Fix issue setting MemberOf property.
     FIXES [#3496](https://github.com/microsoft/Microsoft365DSC/issues/3496)
+* MISC
+  * Improved Update-M365DSCDependencies function to properly install all Microsoft.Graph.*M modules.
+    FIXES [#3454](https://github.com/microsoft/Microsoft365DSC/issues/3454)
 
 # 1.23.719.1